# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchPromptEnhancementItemView": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "revisionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revisionid"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "skipped",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "state"
        ],
        "title": "BatchPromptEnhancementItemView",
        "type": "object"
      },
      "BatchPromptEnhancementRequest": {
        "properties": {
          "includePreviousTaskSummary": {
            "default": false,
            "title": "Includeprevioustasksummary",
            "type": "boolean"
          },
          "includeProjectBackground": {
            "default": false,
            "title": "Includeprojectbackground",
            "type": "boolean"
          },
          "taskIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Taskids",
            "type": "array"
          }
        },
        "title": "BatchPromptEnhancementRequest",
        "type": "object"
      },
      "BatchPromptEnhancementResponse": {
        "properties": {
          "batchId": {
            "title": "Batchid",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchPromptEnhancementItemView"
            },
            "title": "Items",
            "type": "array"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "batchId",
          "state",
          "items"
        ],
        "title": "BatchPromptEnhancementResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectCoverSelection": {
        "properties": {
          "assetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetid"
          },
          "kind": {
            "enum": [
              "auto",
              "asset",
              "video"
            ],
            "title": "Kind",
            "type": "string"
          },
          "resultId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultid"
          },
          "seconds": {
            "default": 0,
            "minimum": 0.0,
            "title": "Seconds",
            "type": "number"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ProjectCoverSelection",
        "type": "object"
      },
...
      "ProjectPatchRequest": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectCoverSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "useDescriptionForAiPrompt": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usedescriptionforaiprompt"
          }
        },
        "title": "ProjectPatchRequest",
        "type": "object"
      },
...
      "ProjectSummary": {
        "properties": {
          "assetCount": {
            "title": "Assetcount",
            "type": "integer"
          },
          "completedTaskCount": {
            "default": 0,
            "title": "Completedtaskcount",
            "type": "integer"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverurl"
          },
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "generationWarnings": {
            "items": {
              "type": "string"
            },
            "title": "Generationwarnings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "newResults": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Newresults",
            "type": "array"
          },
          "status": {
            "enum": [
              "idle",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "taskCount": {
            "title": "Taskcount",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "taskCount",
          "assetCount",
          "updatedAt"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
...
      "PromptBatchEligibilityRequest": {
        "properties": {
          "taskIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Taskids",
            "type": "array"
          }
        },
        "title": "PromptBatchEligibilityRequest",
        "type": "object"
      },
      "PromptBatchEligibilityView": {
        "properties": {
          "eligibleTaskIds": {
            "items": {
              "type": "string"
            },
            "title": "Eligibletaskids",
            "type": "array"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/PromptBatchSkippedItem"
            },
            "title": "Skipped",
            "type": "array"
          }
        },
        "required": [
          "eligibleTaskIds",
          "skipped"
        ],
        "title": "PromptBatchEligibilityView",
        "type": "object"
      },
...
      "PromptBatchSkippedItem": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "enum": [
              "busy",
              "empty-prompt"
            ],
            "title": "Reason",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "reason",
          "message"
        ],
        "title": "PromptBatchSkippedItem",
        "type": "object"
      },
...
      "PromptReviewItemView": {
        "properties": {
          "approvedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approvedat"
          },
          "approvedRevision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approvedrevision"
          },
          "promptReviewStatus": {
            "enum": [
              "not_ready",
              "pending_review",
              "approved"
            ],
            "title": "Promptreviewstatus",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "promptReviewStatus"
        ],
        "title": "PromptReviewItemView",
        "type": "object"
      },
      "PromptReviewStateView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PromptReviewItemView"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PromptReviewStateView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoBatchEligibilityView": {
        "properties": {
          "eligibleTaskIds": {
            "items": {
              "type": "string"
            },
            "title": "Eligibletaskids",
            "type": "array"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/VideoBatchSkippedItem"
            },
            "title": "Skipped",
            "type": "array"
          },
          "warnings": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "eligibleTaskIds",
          "skipped"
        ],
        "title": "VideoBatchEligibilityView",
        "type": "object"
      },
      "VideoBatchRequest": {
        "properties": {
          "taskIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Taskids",
            "type": "array"
          }
        },
        "title": "VideoBatchRequest",
        "type": "object"
      },
      "VideoBatchResponse": {
        "properties": {
          "batchId": {
            "title": "Batchid",
            "type": "string"
          },
          "eligibleTaskIds": {
            "items": {
              "type": "string"
            },
            "title": "Eligibletaskids",
            "type": "array"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/VideoBatchSkippedItem"
            },
            "title": "Skipped",
            "type": "array"
          },
          "warnings": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "eligibleTaskIds",
          "skipped",
          "batchId"
        ],
        "title": "VideoBatchResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ShotMill",
    "version": "0.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}": {
      "patch": {
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/prompt-enhancement-batches": {
      "post": {
        "operationId": "create_prompt_enhancement_batch_api_v1_projects__project_id__prompt_enhancement_batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPromptEnhancementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPromptEnhancementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Prompt Enhancement Batch",
        "tags": [
          "batch-review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/prompt-enhancement-batches/eligibility": {
      "post": {
        "operationId": "check_prompt_batch_eligibility_api_v1_projects__project_id__prompt_enhancement_batches_eligibility_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptBatchEligibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptBatchEligibilityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Prompt Batch Eligibility",
        "tags": [
          "batch-review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/prompt-enhancement-batches/{batch_id}": {
      "get": {
        "operationId": "get_prompt_enhancement_batch_api_v1_projects__project_id__prompt_enhancement_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPromptEnhancementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prompt Enhancement Batch",
        "tags": [
          "batch-review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/prompt-enhancement-batches/{batch_id}/cancel": {
      "post": {
        "operationId": "cancel_prompt_enhancement_batch_api_v1_projects__project_id__prompt_enhancement_batches__batch_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPromptEnhancementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Prompt Enhancement Batch",
        "tags": [
          "batch-review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/prompt-enhancement-batches/{batch_id}/retry-failed": {
      "post": {
        "operationId": "retry_failed_prompt_enhancement_batch_api_v1_projects__project_id__prompt_enhancement_batches__batch_id__retry_failed_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPromptEnhancementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Failed Prompt Enhancement Batch",
        "tags": [
          "batch-review"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/prompt-review-state": {
      "get": {
        "operationId": "get_prompt_review_state_api_v1_projects__project_id__prompt_review_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptReviewStateView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prompt Review State",
        "tags": [
          "batch-review"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/tasks/{task_id}/prompt-review": {
      "post": {
        "operationId": "approve_prompt_api_v1_projects__project_id__tasks__task_id__prompt_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptReviewItemView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Prompt",
        "tags": [
          "batch-review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/tasks/{task_id}/prompt-review/approve": {
      "post": {
        "operationId": "approve_prompt_api_v1_projects__project_id__tasks__task_id__prompt_review_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptReviewItemView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Prompt",
        "tags": [
          "batch-review"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/video-generation-batches": {
      "post": {
        "operationId": "create_video_generation_batch_api_v1_projects__project_id__video_generation_batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Video Generation Batch",
        "tags": [
          "batch-review"
        ]
      }
    },
    "/api/v1/projects/{project_id}/video-generation-batches/eligibility": {
      "post": {
        "operationId": "check_video_batch_eligibility_api_v1_projects__project_id__video_generation_batches_eligibility_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatchEligibilityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Video Batch Eligibility",
        "tags": [
          "batch-review"
        ]
      }
    },
...
  }
}